[FEAT] Memo struct 파일 분리 및 MemoViewModel 생성 #23
