+++ CommonPlant/CommonPlant/Screen/MyPlant/Model/MemoModel.swift
@@ -1,0 +1,16 @@
+//
+//  MemoModel.swift
+//  CommonPlant
+//
+//  Created by 아라 on 2023/08/29.
+//
+
+import Foundation
+
+struct Memo {
+    let userNickName: String
+    let userImgURL: String
+    let content: String
+    let imgURL: String?
+    let createdAt: String
+}
